+++ src/main/java/com/iexec/commons/poco/chain/ChainDataset.java
@@ -23,6 +23,6 @@ import lombok.Value;
 @Builder
 public class ChainDataset {
     String chainDatasetId;
-    String uri;
+    String multiaddr;
     String checksum;
 }