+++ src/main/java/com/iexec/commons/poco/encoding/AccessorsEncoder.java
@@ -50,4 +50,14 @@ public class AccessorsEncoder {
      */
     public static final String OWNER_SELECTOR = "0x8da5cb5b";
 
+    // app
+    public static final String M_APPCHECKSUM_SELECTOR = "0x84aaf12e";
+    public static final String M_APPMRENCLAVE_SELECTOR = "0xe30d26a8";
+    public static final String M_APPMULTIADDR_SELECTOR = "0x39e75d45";
+    public static final String M_APPTYPE_SELECTOR = "0xf8c2ceb3";
+
+    // dataset
+    public static final String M_DATASETCHECKSUM_SELECTOR = "0x1ba99d7e";
+    public static final String M_DATASETMULTIADDR_SELECTOR = "0xa61ca6c5";
+
 }