+++ src/test/java/com/iexec/commons/poco/itest/IexecHubTestService.java
@@ -42,8 +42,8 @@ public class IexecHubTestService extends IexecHubAbstractService {
     static final BigInteger GAS_LIMIT = BigInteger.valueOf(1_000_000L);
     static final String IEXEC_HUB_ADDRESS = "0xc4b11f41746D3Ad8504da5B383E1aB9aa969AbC7";
 
-    private static final String ASSET_MULTI_ADDRESS = "multiAddress";
-    private static final String ASSET_CHECKSUM = Numeric.toHexStringNoPrefix(new byte[32]);
+    static final String ASSET_MULTI_ADDRESS = "multiAddress";
+    static final String ASSET_CHECKSUM = Numeric.toHexStringNoPrefix(new byte[32]);
 
     private static final String APP_REGISTRY_SELECTOR = "0x45b637a9";
     private static final String DATASET_REGISTRY_SELECTOR = "0xb1b11d2c";